requirements.txt
streamlit
pandas
numpy
openai
google-genai